+++ src/core/ndid.js
@@ -20,9 +20,10 @@
  *
  */
 
-import * as utils from '../utils';
 import * as tendermint from '../tendermint';
+import { validateKeyType } from './common';
 import logger from '../logger';
+import * as utils from '../utils';
 import { callbackToClient } from '../utils/callback';
 
 import CustomError from '../error/custom_error';
@@ -30,7 +31,12 @@ import { getErrorObjectForClient } from '../error/helpers';
 
 let init = false;
 
-export async function initNDID({ public_key, master_public_key }) {
+export async function initNDID({
+  public_key,
+  public_key_type,
+  master_public_key,
+  master_public_key_type,
+}) {
   if (init) {
     logger.error({
       message: 'NDID is already exist',
@@ -38,6 +44,14 @@ export async function initNDID({ public_key, master_public_key }) {
     return false;
   }
   try {
+    if (public_key != null) {
+      validateKeyType(public_key, public_key_type);
+    }
+
+    if (master_public_key != null) {
+      validateKeyType(master_public_key, master_public_key_type);
+    }
+
     await tendermint.transact(
       'InitNDID',
       { node_id: 'ndid1', public_key, master_public_key },
@@ -112,6 +126,14 @@ export async function reduceNodeToken(data) {
 
 export async function registerNode(data, { synchronous = false } = {}) {
   try {
+    if (data.public_key != null) {
+      validateKeyType(data.public_key, data.public_key_type);
+    }
+
+    if (data.master_public_key != null) {
+      validateKeyType(data.master_public_key, data.master_public_key_type);
+    }
+
     if (synchronous) {
       await registerNodeInternalAsync(...arguments);
     } else {
@@ -209,8 +231,6 @@ async function updateNodeInternalAsync(data, { synchronous = false } = {}) {
     // role,
     max_ial,
     max_aal,
-    public_key,
-    master_public_key,
   } = data;
 
   try {