+++ src/error/type.js
@@ -316,6 +316,21 @@ export default {
     message: 'Invalid data type in AS data response. Expected string.',
     clientError: true,
   },
+  INVALID_KEY_FORMAT: {
+    code: 20036,
+    message: 'Invalid key format',
+    clientError: true,
+  },
+  UNSUPPORTED_KEY_TYPE: {
+    code: 20037,
+    message: 'Unsupported key type. Only RSA is allowed.',
+    clientError: true,
+  },
+  MISMATCHED_KEY_TYPE: {
+    code: 20038,
+    message: 'Key type mismatched. Provided key type does not match with given key.',
+    clientError: true,
+  },
 
   // Errors return from ABCI app
   // Server errors