+++ src/routes/v1/dpki.js
@@ -35,11 +35,11 @@ router.post('/node/create', validateBody, async (req, res, next) => {
       node_id,
       node_name,
       node_key,
-      //node_key_type,
-      //node_key_method,
+      node_key_type,
+      // node_sign_method,
       node_master_key,
-      //node_master_key_type,
-      //node_master_key_method,
+      node_master_key_type,
+      // node_master_sign_method,
       role,
       max_aal,
       max_ial,
@@ -50,7 +50,9 @@ router.post('/node/create', validateBody, async (req, res, next) => {
         node_id,
         node_name,
         public_key: node_key,
+        public_key_type: node_key_type,
         master_public_key: node_master_key,
+        master_public_key_type: node_master_key_type,
         role,
         max_ial,
         max_aal,
@@ -69,18 +71,20 @@ router.post('/node/update', validateBody, async (req, res, next) => {
     const {
       //node_name,
       node_key,
-      //node_key_type,
-      //node_key_method,
+      node_key_type,
+      // node_sign_method,
       node_master_key,
-      //node_master_key_type,
-      //node_master_key_method,
+      node_master_key_type,
+      // node_master_sign_method,
     } = req.body;
 
     //should we allow organization to update their node's name?
     let result = await dpki.updateNode(
       {
         public_key: node_key,
+        public_key_type: node_key_type,
         master_public_key: node_master_key,
+        master_public_key_type: node_master_key_type,
       },
       { synchronous: true }
     );